+++ apps/decodex/src/agent/app_server/tests.rs
@@ -1,5 +1,7 @@
 mod archive;
 mod dynamic_tools;
+mod phase_goal_runtime;
+mod phase_goal_support;
 mod phase_goal_tests;
 mod preflight;
 mod recorder;
@@ -9,7 +11,7 @@ mod schema_tests;
 
 use std::{
 	cell::RefCell,
-	env, fs,
+	fs,
 	os::unix::fs::PermissionsExt,
 	path::{Path, PathBuf},
 	time::Duration,
@@ -24,11 +26,10 @@ use crate::{
 			APP_SERVER_REQUIRED_CLIENT_NOTIFICATIONS, APP_SERVER_REQUIRED_CLIENT_REQUESTS,
 			APP_SERVER_REQUIRED_SERVER_NOTIFICATIONS, APP_SERVER_REQUIRED_SERVER_REQUESTS,
 			APP_SERVER_SCHEMA_REQUIRED_MARKERS, AppServerCapabilityPreflightFailure,
-			AppServerCapabilityPreflightReport, AppServerPhaseGoalFailure, AppServerRunResult,
-			AppServerThreadArchiveOutcome, AppServerThreadArchiveRequest, AppServerTurnFailure,
-			CommandExecHealthCheck, CommandExecResponse, EffectiveThreadConfig, InitializeResponse,
-			ModelProviderCapabilitiesReadResponse, PhaseGoalController, PhaseGoalKind,
-			PhaseGoalSpec, PhaseGoalTransition, PluginListResponse, ProbeDynamicToolHandler,
+			AppServerCapabilityPreflightReport, AppServerThreadArchiveOutcome,
+			AppServerThreadArchiveRequest, AppServerTurnFailure, CommandExecHealthCheck,
+			CommandExecResponse, EffectiveThreadConfig, InitializeResponse,
+			ModelProviderCapabilitiesReadResponse, PluginListResponse, ProbeDynamicToolHandler,
 			REQUEST_TIMEOUT, RunRecorder, RuntimeConfigSummary, SkillsListResponse,
 			TurnContinuationGuard, archive_app_server_thread_after_success,
 			classify_turn_completion, continuation_boundary_reached, execute_app_server_run,
@@ -57,172 +58,12 @@ use crate::{
 	run_control::{
 		LaneControlSteerRequest, LaneControlSteerRequestInput, LaneControlSteerResponse,
 	},
-	state::StateStore,
-	test_support::TestEnvVarGuard,
 };
-
-const PHASE_GOAL_FAKE_CODEX_SCRIPT_TEMPLATE: &str = r#"#!/usr/bin/env python3
-import json
-import os
-import sys
-
-TURN_OUTPUTS = __TURN_OUTPUTS__
-GOAL_STATUSES = __GOAL_STATUSES__
-UNSUPPORTED_GOAL_METHODS = __UNSUPPORTED_GOAL_METHODS__
-MISMATCH_NOTIFICATION_TURN_IDS = __MISMATCH_NOTIFICATION_TURN_IDS__
-
-goal = {
-    "objective": "",
-    "tokenBudget": None,
-}
-turn_count = 0
-goal_get_count = 0
-
-def send(value):
-    print(json.dumps(value), flush=True)
-
-def reply(message_id, result):
-    send({"id": message_id, "result": result})
-
-def method_not_found(message_id, method):
-    send({"id": message_id, "error": {
-        "code": -32601,
-        "message": "Method not found: " + str(method),
-    }})
-
-def unsupported_goal_method(method):
-    return method in UNSUPPORTED_GOAL_METHODS
-
-def goal_payload(status):
-    return {
-        "createdAt": 1,
-        "objective": goal["objective"],
-        "status": status,
-        "threadId": "thread-1",
-        "timeUsedSeconds": 0,
-        "tokenBudget": goal["tokenBudget"],
-        "tokensUsed": 0,
-        "updatedAt": 1,
-    }
-
-for line in sys.stdin:
-    message = json.loads(line)
-    method = message.get("method")
-    message_id = message.get("id")
-    params = message.get("params") or {}
-
-    if method == "initialize":
-        reply(message_id, {
-            "userAgent": "codex-cli 0.136.0",
-            "codexHome": os.environ["CODEX_HOME"],
-            "platformFamily": "unix",
-            "platformOs": "macos",
-        })
-    elif method == "initialized":
-        continue
-    elif method == "config/read":
-        reply(message_id, {"config": {
-            "model": "gpt-5.5",
-            "model_provider": "openai",
-            "approval_policy": {"type": "never"},
-            "sandbox_mode": {"type": "dangerFullAccess"},
-        }})
-    elif method == "model/list":
-        reply(message_id, {"data": [{
-            "id": "gpt-5.5",
-            "model": "gpt-5.5",
-            "displayName": "GPT-5.5",
-            "isDefault": True,
-            "hidden": False,
-        }], "nextCursor": None})
-    elif method == "modelProvider/capabilities/read":
-        reply(message_id, {"imageGeneration": True, "namespaceTools": True, "webSearch": True})
-    elif method == "skills/list":
-        cwd = params.get("cwds", [""])[0]
-        reply(message_id, {"data": [{"cwd": cwd, "errors": [], "skills": [{
-            "enabled": True,
-            "name": "fake-skill",
-            "scope": "user",
-        }]}]})
-    elif method == "plugin/list":
-        reply(message_id, {"marketplaces": [{"name": "fake", "plugins": [{
-            "enabled": True,
-            "id": "fake-plugin",
-            "installed": True,
-            "name": "Fake Plugin",
-        }]}], "marketplaceLoadErrors": []})
-    elif method == "mcpServerStatus/list":
-        reply(message_id, {"data": [], "nextCursor": None})
-    elif method == "thread/start":
-        reply(message_id, {
-            "thread": {"id": "thread-1"},
-            "model": "gpt-5.5",
-            "modelProvider": "openai",
-            "serviceTier": None,
-            "cwd": params.get("cwd"),
-            "instructionSources": [],
-            "approvalPolicy": {"type": "never"},
-            "approvalsReviewer": "user",
-            "sandbox": {"type": "dangerFullAccess"},
-            "reasoningEffort": None,
-        })
-    elif method == "thread/goal/set":
-        if unsupported_goal_method(method):
-            method_not_found(message_id, method)
-        else:
-            goal["objective"] = params.get("objective") or ""
-            goal["tokenBudget"] = params.get("tokenBudget")
-            reply(message_id, {"goal": goal_payload("active")})
-    elif method == "thread/goal/get":
-        if unsupported_goal_method(method):
-            method_not_found(message_id, method)
-        else:
-            if GOAL_STATUSES:
-                status = GOAL_STATUSES[min(goal_get_count, len(GOAL_STATUSES) - 1)]
-            else:
-                status = "active"
-            goal_get_count += 1
-            reply(message_id, {"goal": None if status == "none" else goal_payload(status)})
-    elif method == "thread/goal/clear":
-        if unsupported_goal_method(method):
-            method_not_found(message_id, method)
-        else:
-            reply(message_id, {"cleared": True})
-    elif method == "turn/start":
-        turn_count += 1
-        turn_id = "turn-" + str(turn_count)
-        notification_turn_id = "notification-" + turn_id if MISMATCH_NOTIFICATION_TURN_IDS else turn_id
-        if TURN_OUTPUTS:
-            output = TURN_OUTPUTS[min(turn_count - 1, len(TURN_OUTPUTS) - 1)]
-        else:
-            output = "DONE"
-        reply(message_id, {"turn": {"id": turn_id, "status": "running", "error": None}})
-        send({"method": "thread/status/changed", "params": {
-            "threadId": "thread-1",
-            "status": {"type": "active", "activeFlags": []},
-        }})
-        send({"method": "turn/started", "params": {
-            "threadId": "thread-1",
-            "turn": {"id": notification_turn_id, "status": "running", "error": None},
-        }})
-        if not unsupported_goal_method("thread/goal/updated"):
-            send({"method": "thread/goal/updated", "params": {
-                "threadId": "thread-1",
-                "turnId": notification_turn_id,
-                "goal": goal_payload("active"),
-            }})
-        send({"method": "item/completed", "params": {
-            "threadId": "thread-1",
-            "turnId": notification_turn_id,
-            "item": {"type": "agentMessage", "text": output},
-        }})
-        send({"method": "turn/completed", "params": {
-            "threadId": "thread-1",
-            "turn": {"id": notification_turn_id, "status": "completed", "error": None},
-        }})
-    else:
-        method_not_found(message_id, method)
-"#;
+use phase_goal_support::{
+	ContinueTokenCompletionHandler, TerminalTokenCompletionHandler, TestPhaseGoalController,
+	execute_phase_goal_fake_app_server, phase_goal_fake_codex_script,
+	phase_goal_fake_codex_script_with_notification_turn_mismatch, private_phase_goal_events,
+};
 
 struct RejectingCompletionHandler;
 impl DynamicToolHandler for RejectingCompletionHandler {
@@ -410,89 +251,6 @@ struct LiveResumeBoundaryGuard;
 impl TurnContinuationGuard for LiveResumeBoundaryGuard {
 	fn should_continue_turn(&self, _turn_count: u32) -> Result<bool> {
 		Ok(false)
-	}
-}
-
-struct ContinueTokenCompletionHandler;
-impl DynamicToolHandler for ContinueTokenCompletionHandler {
-	fn tool_specs(&self) -> Vec<DynamicToolSpec> {
-		Vec::new()
-	}
-
-	fn handle_call(&self, _tool_name: &str, _arguments: Value) -> DynamicToolCallResponse {
-		DynamicToolCallResponse::failure(String::from("unused"))
-	}
-
-	fn classify_turn_completion(&self, final_output: &str) -> Result<TurnCompletionStatus> {
-		Ok(if final_output.trim() == "CONTINUE" {
-			TurnCompletionStatus::Continue
-		} else {
-			TurnCompletionStatus::Complete
-		})
-	}
-}
-
-#[derive(Default)]
-struct TerminalTokenCompletionHandler {
-	terminal_seen: RefCell<bool>,
-}
-impl DynamicToolHandler for TerminalTokenCompletionHandler {
-	fn tool_specs(&self) -> Vec<DynamicToolSpec> {
-		Vec::new()
-	}
-
-	fn handle_call(&self, _tool_name: &str, _arguments: Value) -> DynamicToolCallResponse {
-		DynamicToolCallResponse::failure(String::from("unused"))
-	}
-
-	fn classify_turn_completion(&self, final_output: &str) -> Result<TurnCompletionStatus> {
-		Ok(match final_output.trim() {
-			"CONTINUE" => TurnCompletionStatus::Continue,
-			"TERMINAL" => {
-				self.terminal_seen.replace(true);
-
-				TurnCompletionStatus::Complete
-			},
-			_ => TurnCompletionStatus::Complete,
-		})
-	}
-
-	fn has_terminal_completion_signal(&self) -> bool {
-		*self.terminal_seen.borrow()
-	}
-}
-
-struct TestPhaseGoalController {
-	initial_phase: PhaseGoalKind,
-}
-impl TestPhaseGoalController {
-	fn new(initial_phase: PhaseGoalKind) -> Self {
-		Self { initial_phase }
-	}
-}
-
-impl PhaseGoalController for TestPhaseGoalController {
-	fn initial_phase_goal(&self) -> Result<Option<PhaseGoalSpec>> {
-		Ok(Some(PhaseGoalSpec::new(self.initial_phase, "test phase goal", None)))
-	}
-
-	fn phase_goal_completed(&self, phase: PhaseGoalKind) -> Result<PhaseGoalTransition> {
-		Ok(match phase {
-			PhaseGoalKind::RepairAcceptedReviewFindings =>
-				PhaseGoalTransition::Continue(PhaseGoalSpec::new(
-					PhaseGoalKind::ReviewRepairEvidence,
-					"prepare review repair evidence",
-					None,
-				)),
-			PhaseGoalKind::ImplementToValidationReady | PhaseGoalKind::RepairValidationFailures =>
-				PhaseGoalTransition::Continue(PhaseGoalSpec::new(
-					PhaseGoalKind::HandoffEvidence,
-					"prepare handoff evidence",
-					None,
-				)),
-			PhaseGoalKind::ReviewRepairEvidence | PhaseGoalKind::HandoffEvidence =>
-				PhaseGoalTransition::CompleteRun,
-		})
 	}
 }
 
@@ -700,294 +458,4 @@ fn slow_thread_start_fake_codex_script() -> String {
 		"    elif method == \"thread/start\":\n        cwd = params.get(\"cwd\")",
 		"    elif method == \"thread/start\":\n        import time\n        time.sleep(6)\n        cwd = params.get(\"cwd\")",
 	)
-}
-
-fn phase_goal_fake_codex_script(
-	turn_outputs: &[&str],
-	goal_statuses: &[&str],
-	unsupported_goal_methods: &[&str],
-) -> String {
-	phase_goal_fake_codex_script_with_notification_turn_mismatch(
-		turn_outputs,
-		goal_statuses,
-		unsupported_goal_methods,
-		false,
-	)
-}
-
-fn phase_goal_fake_codex_script_with_notification_turn_mismatch(
-	turn_outputs: &[&str],
-	goal_statuses: &[&str],
-	unsupported_goal_methods: &[&str],
-	mismatch_notification_turn_ids: bool,
-) -> String {
-	let outputs_json = serde_json::to_string(turn_outputs).expect("turn outputs should serialize");
-	let statuses_json =
-		serde_json::to_string(goal_statuses).expect("goal statuses should serialize");
-	let unsupported_goal =
-		serde_json::to_string(unsupported_goal_methods).expect("methods should serialize");
-	let mismatch_turn_ids = if mismatch_notification_turn_ids { "True" } else { "False" };
-
-	PHASE_GOAL_FAKE_CODEX_SCRIPT_TEMPLATE
-		.replace("__TURN_OUTPUTS__", &outputs_json)
-		.replace("__GOAL_STATUSES__", &statuses_json)
-		.replace("__UNSUPPORTED_GOAL_METHODS__", &unsupported_goal)
-		.replace("__MISMATCH_NOTIFICATION_TURN_IDS__", mismatch_turn_ids)
-}
-
-fn execute_phase_goal_fake_app_server<'a, F>(
-	script: String,
-	configure: F,
-) -> (Result<AppServerRunResult>, StateStore)
-where
-	F: FnOnce(&mut super::AppServerRunRequest<'a>),
-{
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let worktree_path = temp_dir.path().join("worktree");
-	let fake_bin_dir = install_fake_codex_script(&temp_dir, &script);
-	let path_env = env::var("PATH").unwrap_or_default();
-	let _path_guard =
-		TestEnvVarGuard::set("PATH", &format!("{}:{path_env}", fake_bin_dir.display()));
-	let state_store = StateStore::open_in_memory().expect("state store should open");
-	let mut request = minimal_run_request();
-
-	fs::create_dir_all(&worktree_path).expect("worktree directory should create");
-
-	request.run_id = String::from("phase-goal-run");
-	request.issue_id = String::from("phase-goal-issue");
-	request.cwd = worktree_path.display().to_string();
-	request.timeout = Duration::from_secs(5);
-
-	configure(&mut request);
-
-	let result = super::execute_app_server_run(&request, &state_store);
-
-	(result, state_store)
-}
-
-fn private_phase_goal_events(state_store: &StateStore, event_type: &str) -> Vec<Value> {
-	state_store
-		.list_private_execution_events("test-project", "phase-goal-issue", "phase-goal-run", 1)
-		.expect("private phase goal events should load")
-		.into_iter()
-		.filter(|event| event.event_type() == event_type)
-		.map(|event| event.payload().clone())
-		.collect()
-}
-
-#[test]
-fn phase_goal_set_method_is_required_when_phase_controller_is_present() {
-	let controller = TestPhaseGoalController::new(PhaseGoalKind::ImplementToValidationReady);
-	let script = phase_goal_fake_codex_script(&["DONE"], &[], &["thread/goal/set"]);
-	let (result, _state_store) = execute_phase_goal_fake_app_server(script, |request| {
-		request.phase_goal_controller = Some(&controller);
-	});
-	let error = result.expect_err("missing goal set support should fail immediately");
-	let failure = error
-		.downcast_ref::<AppServerPhaseGoalFailure>()
-		.expect("missing goal support should be a typed phase-goal failure");
-
-	assert_eq!(failure.error_class(), "app_server_phase_goal_unsupported");
-	assert!(error.to_string().contains("thread/goal/set"));
-}
-
-#[test]
-fn phase_goal_get_method_is_required_after_turn_completion() {
-	let controller = TestPhaseGoalController::new(PhaseGoalKind::ImplementToValidationReady);
-	let script = phase_goal_fake_codex_script(&["DONE"], &[], &["thread/goal/get"]);
-	let (result, _state_store) = execute_phase_goal_fake_app_server(script, |request| {
-		request.phase_goal_controller = Some(&controller);
-	});
-	let error = result.expect_err("missing goal get support should fail after the turn");
-	let failure = error
-		.downcast_ref::<AppServerPhaseGoalFailure>()
-		.expect("missing goal support should be a typed phase-goal failure");
-
-	assert_eq!(failure.error_class(), "app_server_phase_goal_unsupported");
-	assert!(error.to_string().contains("thread/goal/get"));
-}
-
-#[test]
-fn phase_goal_complete_runs_validation_transition_before_handoff_goal() {
-	let handler = TerminalTokenCompletionHandler::default();
-	let controller = TestPhaseGoalController::new(PhaseGoalKind::ImplementToValidationReady);
-	let script =
-		phase_goal_fake_codex_script(&["DONE", "TERMINAL"], &["complete", "complete"], &[]);
-	let (result, state_store) = execute_phase_goal_fake_app_server(script, |request| {
-		request.max_turns = 3;
-		request.dynamic_tool_handler = Some(&handler);
-		request.phase_goal_controller = Some(&controller);
-	});
-	let result = result.expect("completed phase goal should advance to handoff evidence goal");
-	let completed_events = private_phase_goal_events(&state_store, "phase_goal_completed");
-	let goal_set_events = private_phase_goal_events(&state_store, "phase_goal_set");
-
-	assert_eq!(result.turn_count, 2);
-	assert_eq!(result.final_output, "TERMINAL");
-	assert_eq!(
-		result.phase_goal_status,
-		Some(super::PhaseGoalRunStatus {
-			phase: PhaseGoalKind::HandoffEvidence,
-			status: String::from("complete"),
-		})
-	);
-	assert_eq!(
-		completed_events.iter().filter_map(|event| event["phase"].as_str()).collect::<Vec<_>>(),
-		vec!["implement_to_validation_ready", "handoff_evidence"]
-	);
-	assert_eq!(goal_set_events.len(), 2);
-	assert_eq!(goal_set_events[1]["phase"], "handoff_evidence");
-}
-
-#[test]
-fn phase_goal_review_repair_validation_transitions_before_repair_evidence_goal() {
-	let handler = TerminalTokenCompletionHandler::default();
-	let controller = TestPhaseGoalController::new(PhaseGoalKind::RepairAcceptedReviewFindings);
-	let script =
-		phase_goal_fake_codex_script(&["DONE", "TERMINAL"], &["complete", "complete"], &[]);
-	let (result, state_store) = execute_phase_goal_fake_app_server(script, |request| {
-		request.max_turns = 3;
-		request.dynamic_tool_handler = Some(&handler);
-		request.phase_goal_controller = Some(&controller);
-	});
-	let result =
-		result.expect("completed review repair goal should advance to repair evidence goal");
-	let completed_events = private_phase_goal_events(&state_store, "phase_goal_completed");
-	let goal_set_events = private_phase_goal_events(&state_store, "phase_goal_set");
-
-	assert_eq!(result.turn_count, 2);
-	assert_eq!(result.final_output, "TERMINAL");
-	assert_eq!(
-		result.phase_goal_status,
-		Some(super::PhaseGoalRunStatus {
-			phase: PhaseGoalKind::ReviewRepairEvidence,
-			status: String::from("complete"),
-		})
-	);
-	assert_eq!(
-		completed_events.iter().filter_map(|event| event["phase"].as_str()).collect::<Vec<_>>(),
-		vec!["repair_accepted_review_findings", "review_repair_evidence"]
-	);
-	assert_eq!(goal_set_events.len(), 2);
-	assert_eq!(goal_set_events[1]["phase"], "review_repair_evidence");
-}
-
-#[test]
-fn phase_goal_completion_accepts_thread_bound_notification_turn_alias() {
-	let handler = TerminalTokenCompletionHandler::default();
-	let controller = TestPhaseGoalController::new(PhaseGoalKind::ImplementToValidationReady);
-	let script = phase_goal_fake_codex_script_with_notification_turn_mismatch(
-		&["DONE", "TERMINAL"],
-		&["complete", "complete"],
-		&[],
-		true,
-	);
-	let (result, state_store) = execute_phase_goal_fake_app_server(script, |request| {
-		request.max_turns = 3;
-		request.dynamic_tool_handler = Some(&handler);
-		request.phase_goal_controller = Some(&controller);
-	});
-	let result = result.expect("thread-bound turn alias should still complete phase goals");
-	let completed_events = private_phase_goal_events(&state_store, "phase_goal_completed");
-	let run_attempt = state_store
-		.run_attempt("phase-goal-run")
-		.expect("run attempt should load")
-		.expect("run attempt should exist");
-
-	assert_eq!(result.turn_count, 2);
-	assert_eq!(result.turn_id, "notification-turn-2");
-	assert_eq!(run_attempt.turn_id(), Some("notification-turn-2"));
-	assert_eq!(
-		completed_events.iter().filter_map(|event| event["phase"].as_str()).collect::<Vec<_>>(),
-		vec!["implement_to_validation_ready", "handoff_evidence"]
-	);
-}
-
-#[test]
-fn open_phase_goal_stops_at_max_turns_without_terminal_signal() {
-	let handler = ContinueTokenCompletionHandler;
-	let controller = TestPhaseGoalController::new(PhaseGoalKind::ImplementToValidationReady);
-	let script = phase_goal_fake_codex_script(&["CONTINUE", "DONE"], &["active", "active"], &[]);
-	let (result, _state_store) = execute_phase_goal_fake_app_server(script, |request| {
-		request.max_turns = 2;
-		request.dynamic_tool_handler = Some(&handler);
-		request.phase_goal_controller = Some(&controller);
-	});
-	let result = result.expect("open phase goal should allow another bounded turn");
-
-	assert_eq!(result.turn_count, 2);
-	assert_eq!(result.turn_id, "turn-2");
-	assert_eq!(result.final_output, "DONE");
-	assert!(result.continuation_pending);
-	assert_eq!(
-		result.phase_goal_status,
-		Some(super::PhaseGoalRunStatus {
-			phase: PhaseGoalKind::ImplementToValidationReady,
-			status: String::from("active"),
-		})
-	);
-}
-
-#[test]
-fn open_phase_goal_stops_at_max_turns_with_continuation_pending() {
-	let handler = ContinueTokenCompletionHandler;
-	let controller = TestPhaseGoalController::new(PhaseGoalKind::ImplementToValidationReady);
-	let script = phase_goal_fake_codex_script(&["CONTINUE"], &["active"], &[]);
-	let (result, _state_store) = execute_phase_goal_fake_app_server(script, |request| {
-		request.max_turns = 1;
-		request.dynamic_tool_handler = Some(&handler);
-		request.phase_goal_controller = Some(&controller);
-	});
-	let result = result.expect("open phase goal should exit cleanly at max_turns");
-
-	assert_eq!(result.turn_count, 1);
-	assert!(result.continuation_pending);
-	assert_eq!(
-		result.phase_goal_status,
-		Some(super::PhaseGoalRunStatus {
-			phase: PhaseGoalKind::ImplementToValidationReady,
-			status: String::from("active"),
-		})
-	);
-}
-
-#[test]
-fn phase_goal_handoff_continue_without_terminal_completion_is_invalid() {
-	let handler = ContinueTokenCompletionHandler;
-	let controller = TestPhaseGoalController::new(PhaseGoalKind::HandoffEvidence);
-	let script = phase_goal_fake_codex_script(&["CONTINUE"], &["complete"], &[]);
-	let (result, _state_store) = execute_phase_goal_fake_app_server(script, |request| {
-		request.max_turns = 2;
-		request.dynamic_tool_handler = Some(&handler);
-		request.phase_goal_controller = Some(&controller);
-	});
-	let error = result.expect_err("handoff goal completion cannot replace terminal path");
-	let failure = error
-		.downcast_ref::<AppServerPhaseGoalFailure>()
-		.expect("missing terminal path should be a typed phase-goal failure");
-
-	assert_eq!(failure.error_class(), "phase_goal_terminal_path_missing");
-	assert!(error.to_string().contains("handoff_evidence"));
-}
-
-#[test]
-fn phase_goal_handoff_final_output_without_terminal_signal_is_invalid() {
-	let handler = ContinueTokenCompletionHandler;
-	let controller = TestPhaseGoalController::new(PhaseGoalKind::ImplementToValidationReady);
-	let script = phase_goal_fake_codex_script(&["DONE", "DONE"], &["complete", "complete"], &[]);
-	let (result, state_store) = execute_phase_goal_fake_app_server(script, |request| {
-		request.max_turns = 3;
-		request.dynamic_tool_handler = Some(&handler);
-		request.phase_goal_controller = Some(&controller);
-	});
-	let error = result.expect_err("handoff goal final output cannot replace terminal path");
-	let failure = error
-		.downcast_ref::<AppServerPhaseGoalFailure>()
-		.expect("missing terminal path should be a typed phase-goal failure");
-	let goal_set_events = private_phase_goal_events(&state_store, "phase_goal_set");
-
-	assert_eq!(failure.error_class(), "phase_goal_terminal_path_missing");
-	assert!(error.to_string().contains("handoff_evidence"));
-	assert_eq!(goal_set_events.len(), 2);
-	assert_eq!(goal_set_events[1]["phase"], "handoff_evidence");
 }
